ci: production-grade pipeline — ruff, mypy, coverage gate, CodeQL, dependabot
Pipeline:
- new 'quality' job: uv lock --check, ruff lint+format, mypy, build
- test job: coverage gate at 85% (currently 87%)
- concurrency cancellation to stop superseded runs
- CodeQL security scanning (push/PR/weekly)
- dependabot for github-actions and python deps (weekly, grouped)
- SECURITY.md with private reporting + sandbox scope notes

Code quality (ruff/mypy findings):
- brain: lambda-in-loop binding (B023) -> functools.partial; typed user_content
- hands: line lengths, ternary
- wake: iterate prediction values; local model binding for type-safety
- audio: float() around np.mean

Tests: 108 -> 120, coverage 79% -> 87%
- tests/test_integration.py: end-to-end flows on real filesystem + real
  dispatch (voice command -> disk, memory across session restart, todo
  lifecycle, sandbox adversarial sweep, malformed-args resilience)
- wake: real-thread lifecycle tests (zombie-stream regression coverage)
- mouth: generation + interrupt-mid-playback
- main: assemble_tools extracted and covered for both branches

+++ src/jarvis/audio.py
@@ -48,7 +48,7 @@ async def record_until_silence(
         if threshold is None:
             calibration_samples.append(amplitude)
             if len(calibration_samples) >= CALIBRATION_CHUNKS:
-                ambient_est = np.mean(calibration_samples)
+                ambient_est = float(np.mean(calibration_samples))
                 threshold = _clamp_threshold(ambient_est * settings.silence_multiplier, settings)
                 print(f"[Audio] Calibrated: ambient={ambient_est:.4f}, threshold={threshold:.4f}")
             return
